Move Reporting/components/VesselReportingList into Vessel/components/VesselSidebar/
