+++ frontend/src/features/Reporting/components/ReportingCard.tsx
@@ -1,13 +1,16 @@
-import { getFrenchOrdinal, getReportingActorLabel } from '@features/Reporting/components/VesselReportingList/utils'
 import { deleteReporting } from '@features/Reporting/useCases/deleteReporting'
 import { reportingIsAnInfractionSuspicion } from '@features/Reporting/utils'
 import { getAlertNameFromType } from '@features/SideWindow/Alert/AlertListAndReportingList/utils'
+import { DeletionConfirmationModal } from '@features/Vessel/components/VesselSidebar/VesselReportingList/DeletionConfirmationModal'
+import {
+  getFrenchOrdinal,
+  getReportingActorLabel
+} from '@features/Vessel/components/VesselSidebar/VesselReportingList/utils'
 import { useMainAppDispatch } from '@hooks/useMainAppDispatch'
 import { Accent, Icon, IconButton, THEME, Link } from '@mtes-mct/monitor-ui'
 import { useMemo, useState } from 'react'
 import styled from 'styled-components'
 
-import { DeletionConfirmationModal } from './VesselReportingList/DeletionConfirmationModal'
 import { getDateTime } from '../../../utils'
 import { ReportingType, ReportingTypeCharacteristics } from '../types'
 import { archiveReporting } from '../useCases/archiveReporting'
