+++ frontend/src/features/Vessel/components/VesselSidebar/VesselReportingList/Summary/InfractionSuspicionSummary.tsx
